# src/client/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "LemonDo.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/reveal-profile": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevealProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevealedProfileResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfAdminUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{id}/roles": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/users/{id}/roles/{roleName}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "roleName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/users/{id}/deactivate": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/users/{id}/reactivate": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/users/{id}/reveal": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevealProtectedDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevealedProtectedDataDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/tasks/{taskId}/reveal-note": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevealTaskNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/audit": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "dateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AuditAction"
            }
          },
          {
            "name": "actorId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "resourceType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfAuditEntryDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminUserDto": {
        "required": [
          "id",
          "email",
          "displayName",
          "roles",
          "isActive",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User's unique identifier.",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "User's email (redacted by default, e.g. \"j***@example.com\")."
          },
          "displayName": {
            "type": "string",
            "description": "User's display name (redacted by default, e.g. \"J***e\")."
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Roles assigned to the user."
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the user account is active (not locked out permanently)."
          },
          "createdAt": {
            "type": "string",
            "description": "When the user account was created.",
            "format": "date-time"
          }
        },
        "description": "Read model for user data in admin views. Protected data fields are redacted by default."
      },
...
      "AssignRoleRequest": {
        "required": [
          "roleName"
        ],
        "type": "object",
        "properties": {
          "roleName": {
            "type": "string"
          }
        },
        "description": "Assigns a role to a user account."
      },
      "AuditAction": {
        "enum": [
          "UserRegistered",
          "UserLoggedIn",
          "UserLoggedOut",
          "RoleAssigned",
          "RoleRemoved",
          "ProtectedDataRevealed",
          "TaskCreated",
          "TaskCompleted",
          "TaskDeleted",
          "UserDeactivated",
          "UserReactivated",
          "ProtectedDataAccessed",
          "SensitiveNoteRevealed",
          "OwnProfileRevealed"
        ],
        "description": "Categorizes security-relevant actions recorded in the audit trail."
      },
      "AuditEntryDto": {
        "required": [
          "id",
          "timestamp",
          "actorId",
          "action",
          "resourceType",
          "resourceId",
          "details",
          "ipAddress"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "actorId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "action": {
            "$ref": "#/components/schemas/AuditAction"
          },
          "resourceType": {
            "type": "string"
          },
          "resourceId": {
            "type": [
              "null",
              "string"
            ]
          },
          "details": {
            "type": [
              "null",
              "string"
            ]
          },
          "ipAddress": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Data transfer object for an audit trail entry."
      },
      "AuthResponse": {
        "required": [
          "accessToken",
          "user"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "description": "Authentication response containing access token and user profile."
      },
...
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "description": "Authenticates a user with email and password credentials."
      },
...
      "PagedResultOfAdminUserDto": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserDto"
            }
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Total number of pages available based on int PagedResult&lt;T&gt;.TotalCount and int PagedResult&lt;T&gt;.PageSize.",
            "format": "int32"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "True if there are more pages after the current int PagedResult&lt;T&gt;.Page."
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "True if there are pages before the current int PagedResult&lt;T&gt;.Page. False when on the first page."
          }
        },
        "description": "Represents a paginated result set."
      },
      "PagedResultOfAuditEntryDto": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEntryDto"
            }
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Total number of pages available based on int PagedResult&lt;T&gt;.TotalCount and int PagedResult&lt;T&gt;.PageSize.",
            "format": "int32"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "True if there are more pages after the current int PagedResult&lt;T&gt;.Page."
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "True if there are pages before the current int PagedResult&lt;T&gt;.Page. False when on the first page."
          }
        },
        "description": "Represents a paginated result set."
      },
...
      "RegisterRequest": {
        "required": [
          "email",
          "password",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        },
        "description": "Registers a new user account with email and password."
      },
...
      "RevealedProfileResponse": {
        "required": [
          "email",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        },
        "description": "Response containing the user's unredacted profile data after password re-authentication."
      },
      "RevealedProtectedDataDto": {
        "required": [
          "email",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        },
        "description": "DTO returned when protected data is revealed."
      },
      "RevealProfileRequest": {
        "required": [
          "password"
        ],
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "description": "Request body for `POST /api/auth/reveal-profile`."
      },
      "RevealProtectedDataRequest": {
        "required": [
          "reason",
          "reasonDetails",
          "comments",
          "password"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          },
          "reasonDetails": {
            "type": [
              "null",
              "string"
            ]
          },
          "comments": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": "string"
          }
        },
        "description": "Reveals a user's protected data via break-the-glass authentication."
      },
      "RevealTaskNoteRequest": {
        "required": [
          "reason",
          "reasonDetails",
          "comments",
          "password"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          },
          "reasonDetails": {
            "type": [
              "null",
              "string"
            ]
          },
          "comments": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": "string"
          }
        },
        "description": "Reveals a task's sensitive note via break-the-glass authentication."
      },
...
      "UserResponse": {
        "required": [
          "id",
          "email",
          "displayName",
          "roles"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "User profile information with redacted personal data."
      },
...
    }
  },
...
}